+++ cmd/wfxctl/cmd/workflow/create/create_test.go
@@ -25,44 +25,75 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestCreateWorkflow_YAML(t *testing.T) {
+func TestCreateWorkflow(t *testing.T) {
 	var body []byte
-
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		body, _ = io.ReadAll(r.Body)
-
 		w.Header().Add("Content-Type", "application/json")
 		w.WriteHeader(http.StatusCreated)
-		_, _ = w.Write([]byte("{}"))
+		b, _ := json.Marshal(dau.DirectWorkflow())
+		_, _ = w.Write(b)
 	}))
-	defer ts.Close()
+	t.Cleanup(ts.Close)
 
 	u, _ := url.Parse(ts.URL)
+
+	tests := []struct {
+		name     string
+		marshal  func() ([]byte, error)
+		expected func(input []byte) []byte
+	}{
+		{
+			name: "YAML",
+			marshal: func() ([]byte, error) {
+				return yaml.Marshal(dau.DirectWorkflow())
+			},
+			expected: func(input []byte) []byte {
+				var wf api.Workflow
+				_ = yaml.Unmarshal(input, &wf)
+				b, _ := json.Marshal(wf)
+				return b
+			},
+		},
+		{
+			name: "JSON",
+			marshal: func() ([]byte, error) {
+				return json.Marshal(dau.DirectWorkflow())
+			},
+			expected: func(input []byte) []byte {
+				var wf api.Workflow
+				_ = json.Unmarshal(input, &wf)
+				b, _ := json.Marshal(wf)
+				return b
+			},
+		},
+	}
+
 	t.Setenv("WFX_MGMT_HOST", u.Hostname())
 	t.Setenv("WFX_MGMT_PORT", u.Port())
 
-	f, err := os.CreateTemp("", "wfx-")
-	require.NoError(t, err)
-	defer func() {
-		_ = f.Close()
-		_ = os.Remove(f.Name())
-	}()
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			f, err := os.CreateTemp("", "wfx-")
+			require.NoError(t, err)
+			t.Cleanup(func() {
+				_ = f.Close()
+				_ = os.Remove(f.Name())
+			})
 
-	b, err := yaml.Marshal(dau.DirectWorkflow())
-	require.NoError(t, err)
-	workflowYaml := string(b)
+			b, err := tc.marshal()
+			require.NoError(t, err)
 
-	_, _ = f.WriteString(workflowYaml)
-	cmd := NewCommand()
-	cmd.SetArgs([]string{f.Name()})
+			_, _ = f.Write(b)
+			cmd := NewCommand()
+			cmd.SetArgs([]string{f.Name()})
 
-	err = cmd.Execute()
-	require.NoError(t, err)
+			err = cmd.Execute()
+			require.NoError(t, err)
 
-	var wf api.Workflow
-	_ = yaml.Unmarshal([]byte(workflowYaml), &wf)
-	b, _ = json.Marshal(wf)
-	assert.JSONEq(t, string(b), string(body))
+			assert.JSONEq(t, string(tc.expected(b)), string(body))
+		})
+	}
 }
 
 func TestReadWorkflows_None(t *testing.T) {